Split models command into new file.
