+++ src/chat/commands/commands.ts
@@ -9,7 +9,8 @@ import { unstashCommand } from './context/unstash'
 import { writeCommand } from './context/write'
 import { exitCommand } from './control/exit'
 import { helpCommand } from './control/help'
-import { modelCommand, modelsCommand } from './control/model'
+import { modelCommand } from './control/model'
+import { modelsCommand } from './control/models'
 import { branchCommand } from './conversation/branch'
 import { clearCommand } from './conversation/clear'
 import { continueCommand } from './conversation/continue'
